+++ tests/testthat/test-dependencies.R
@@ -582,3 +582,19 @@ test_that("dependencies() can handle calls", {
   )
 
 })
+
+test_that("dependencies() detects usages of Junit test reporters", {
+
+  check <- function(document) {
+    file <- renv_scope_tempfile("renv-test-", fileext = ".R")
+    writeLines(document, con = file)
+    deps <- dependencies(file, quiet = TRUE)
+    expect_contains(deps$Package, "xml2")
+  }
+
+  check("JunitReporterMock <- R6::R6Class(\"JunitReporterMock\", inherit = JunitReporter)")
+  check("JunitReporter$new()")
+  check("testthat::test_dir(reporter = JunitReporter)")
+  check("testthat::test_dir(reporter = \"junit\")")
+
+})